@@ -8,7 +8,7 @@
     "clean-start": "npm run clean-install && npm start",
     "watch": "webpack --watch --progress --profile",
     "build": "rimraf dist && webpack --progress --profile --bail",
-    "server": "webpack-dev-server --inline --progress --port 8080",
+    "server": "webpack-dev-server --inline --progress --port 8081",
     "webdriver-update": "webdriver-manager update",
     "webdriver-start": "webdriver-manager start",
     "lint": "tslint --force \"src/**/*.ts\"",
@@ -25,14 +25,14 @@
     "postinstall": "npm run webdriver-update"
   },
   "dependencies": {
-    "@angular/common": "~4.2.6",
-    "@angular/compiler": "~4.2.6",
-    "@angular/core": "~4.2.6",
-    "@angular/forms": "~4.2.6",
-    "@angular/http": "~4.2.6",
-    "@angular/platform-browser": "~4.2.6",
-    "@angular/platform-browser-dynamic": "~4.2.6",
-    "@angular/router": "~4.2.6",
+    "@angular/common": "~4.4.7",
+    "@angular/compiler": "~4.4.7",
+    "@angular/core": "~4.4.7",
+    "@angular/forms": "~4.4.7",
+    "@angular/http": "~4.4.7",
+    "@angular/platform-browser": "~4.4.7",
+    "@angular/platform-browser-dynamic": "~4.4.7",
+    "@angular/router": "~4.4.7",
     "core-js": "^2.4.1",
     "reflect-metadata": "^0.1.3",
     "tslib": "^1.7.1",
@@ -43,7 +43,7 @@
     "@angularclass/hmr": "^1.0.1",
     "@angularclass/hmr-loader": "^3.0.2",
     "@types/jasmine": "^2.5.41",
-    "@types/node": "^6.0.38",
+    "@types/node": "^6.0.114",
     "@types/selenium-webdriver": "2.53.39",
     "angular2-template-loader": "^0.6.0",
     "autoprefixer": "^6.3.2",
@@ -67,7 +67,7 @@
     "karma-remap-istanbul": "0.2.1",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "2.0.2",
-    "node-sass": "^4.5.0",
+    "node-sass": "^4.9.2",
     "null-loader": "0.1.1",
     "postcss-loader": "^1.1.0",
     "protractor": "^4.0.10",
