@@ -1,5 +1,5 @@
 exports.config = {
-  baseUrl: 'http://localhost:8080/',
+  baseUrl: 'http://localhost:8081/',
 
   specs: [
     'src/**/*.e2e-spec.js'
