@@ -1,5 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
 
+import { DateService } from '../../shared/services/date.service';
+// import { TasksService } from '../../shared/services/tasks.service';
+
 import { Day } from '../../shared/classes/day';
 
 @Component({
@@ -10,9 +13,14 @@ import { Day } from '../../shared/classes/day';
 export class WorkDayComponent implements OnInit {
     @Input() day: Day;
 
-    constructor() { }
+    constructor(private dateService: DateService) { }
 
     ngOnInit() {
+    }
+
+    setSelectedDay() {
+        console.log('set selected day' + this.day);
+        this.dateService.setSelectedDay(this.day);
     }
 
 }
