@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
+import { TasksService } from '../../shared/services/tasks.service';
+// import { DateService } from '../../shared/services/date.service';
+
+
 @Component({
   selector: 'my-daily-stats',
   templateUrl: './daily-stats.component.html',
@@ -7,7 +11,7 @@ import { Component, OnInit } from '@angular/core';
 })
 export class DailyStatsComponent implements OnInit {
 
-  constructor() { }
+  constructor(public tasksService: TasksService) { }
 
   ngOnInit() {
   }
